requirements.txt
fastapi
uvicorn
snowflake-connector-python
python-dotenv
